Update/badge cards creation flow (#86)
* Added new typechain types

* Updated Badge cards

* ENS support success

* Made UI changes

* Added animation between draft to loading

* save game

* Added basic animation for receipt view

* Added UI to display not enough ETH

* Fixed top page section error

+++ typechain/factories/IBadgeRegistry__factory.ts
@@ -13,9 +13,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint256",
+        internalType: "uint8",
         name: "level",
-        type: "uint256",
+        type: "uint8",
       },
     ],
     name: "getBadgePrice",
